cart0708

+++ app/src/main/java/com/example/myapplication2222/Product.java
@@ -4,10 +4,10 @@ public class Product {
 
     private String id; // 상품 ID
     private String name; // 상품 이름
-    private double price; // 상품 가격
+    private int price; // 상품 가격
     private int quantity; // 상품 수량
 
-    public Product(String id, String name, double price, int quantity) {
+    public Product(String id, String name, int price, int quantity) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -22,7 +22,7 @@ public class Product {
         return name;
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
